+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Unreleased
   - Added benchmark result summary to CI runs
 - Fixed spurious maps file path creation for low addresses as part of
   normalization/symbolization
+- Introduced `blazecli` command line interface for the library
 - Introduced `helper` module exposing `read_elf_build_id` function